+++ extraprecache.cpp
@@ -5,9 +5,9 @@
 #include "extraprecache.h"
 
 model_t* g_ExtraPreacheModel[MAX_EXTRA_PRECACHENUM];
-int iExtraPrecaheSize = 0;
+size_t iExtraPrecaheSize = 0;
 int GetExtraModelIndex(char* path) {
-	for (int i = 0; i < iExtraPrecaheSize; i++) {
+	for (size_t i = 0; i < iExtraPrecaheSize; i++) {
 		if (!g_ExtraPreacheModel[i])
 			continue;
 		if (!strcmp(path, g_ExtraPreacheModel[i]->name))
