+++ itemhighlight.cpp
@@ -36,7 +36,7 @@ void CHudItemHighLight::Reset(){
 	m_mapHighLightTable.clear();
 	m_mapToBeDraw.clear();
 	m_mapRestoredTent.clear();
-	for (int i = 0; i < (int)aryHighLightList.size(); i++) {
+	for (size_t i = 0; i < (int)aryHighLightList.size(); i++) {
 		aryHighLightList[i]->Index = gEngfuncs.pEventAPI->EV_FindModelIndex(aryHighLightList[i]->Name);
 		if (aryHighLightList[i]->Index > 0)
 			m_mapHighLightTable[aryHighLightList[i]->Index] = aryHighLightList[i];
